blog/blog/requirements.txt
asgiref==3.2.6
confusable-homoglyphs==3.2.0
Django==3.0.4
django-crispy-forms==1.9.0
django-registration==3.1
django-widget-tweaks==1.4.8
Pillow==7.1.1
pytz==2019.3
setuptools==46.1.1
sqlparse==0.3.1
wheel==0.34.2
whitenoise